(feat) Tried to save user email to firebase.
Not working, yet. Says helper function is undefined. The idea is to save the user email as an ID to the database, independent of the simple authentication system provided by firebase. Later, when the user updates their profile, the user will be looked up by their email (the id) and the profile info will be added.